Reoprganize for installability (#6)
Reorganize for installability

In order to make stackmon/cloudmon easy installable it is required to
put ansible playbooks and roles either into the collection or place it
inside the package. Collection is a good thing, however in this case it
dramatically reduces ease of development (at least for the initial
phase). For that reason place it inside the project locally with
eventuall splitting out into the collection later. This will require
starting using ansible-builder for prreparing execution container.

Reviewed-by: OpenTelekomCloud Bot